模範回答を確認した。 ・模範回答と作り方が大きく違うので、微修正しかしていない。 ・全表示機能はあるが、残しておく。

+++ app/controllers/line_bot_controller.rb
@@ -21,9 +21,9 @@ class LineBotController < ApplicationController
       if (event.type === Line::Bot::Event::MessageType::Text)
         # LINE からテキストが送信されたときの処理を記述する
         # Task10で追記した
-        message = event["message"]["text"]
-        userid = event["source"]["userId"] #Sp2で追加。
-        message = message.chomp #不要な改行をなくすメソッド（複数個に弱い）
+        message = event["message"]["text"] #lineで入力したメッセージ
+        userid = event["source"]["userId"] #Sp2で追加。ユーザ情報
+        message = message.chomp #不要な改行をなくすメソッド（複数個は無理）
         # binding.pry
 
         # 送信されたメッセージをデータベースに保存するコードを書こう
@@ -53,39 +53,34 @@ class LineBotController < ApplicationController
   def ifmessage(message, userid)
     # binding.pry
     case message
-      when "一覧表示"
-        # 模範回答を元に自分なりに作り直した。
+      when "一覧" #自ユーザの登録したタスク
         # binding.pry
-        list = Task.all.where(user: userid).map.with_index do |task, i| #Sp2で修正した。
-          "#{i+1}: #{task.title}"
-          # "#{task.id}: #{task.title}"
+        list = Task.all.where(user: userid).map.with_index(1) do |task, i| #Sp2で修正した。
+          "#{i}: #{task.title}"
         end.join("\n")
-        # 色々やったけど、each文の使い方として間違ってる。
-        # list2 = Task.all.where(user: userid).pluck(:title) # これで特定のuserの配列を出力
-        # list2.each_with_index do |task, i|
-        #   binding.pry
-        #   test = "#{i+1}：#{task}\n"
-        #   test2 = test+test2
-        # #   puts "#{i}：#{task}"
-        # end
         @linemsg = "【今までに登録したタスク】\n#{list}"
-    	  # tasks.map { |task| "#{index}: #{task.body}" }.join("\n")
-      when /削除\d/
+      when /削除\d/ #特定のタスクを削除する。
         # binding.pry
-        id = message.gsub(/削除/, "").strip.to_i #"削除"と余計な空白を消して残った文字列を数値化
-        task = Task.find(id) #整数化されたidでTaskを出力する
-        task.destroy! # タスクの削除
+        delete = Task.all.where(user: userid) #自ユーザのタスクをHashでもらう[task1,task2...]（各taskにレコードがある）
+        id = message.gsub(/削除/, "").strip.to_i - 1 #"削除"と余計な空白を消して残った文字列を数値化
+        # binding.pry
+        task = delete[id] #整数化されたidでTaskを出力する
+        task.destroy! #タスクの削除
         @linemsg = "タスク #{id}: 「#{task.title}」 を削除しました！"
-      when "削除"
+      when "削除" #自ユーザの最初のタスクを消す
         # binding.pry
-        # これだと最初のタスクしか消せない。
-        text = Task.where(user: userid).first.title #削除されるタスクのタイトル
-        Task.first.destroy #タスクの削除
-        @linemsg = "bomb!\n\"#{text}\"\nは削除されました！！"
-      when "全削除"
-        Task.destroy_all #タスクの削除
-        @linemsg = "bomb!\n全てのタスクは削除されました！"
-      else
+        deltitle = Task.where(user: userid).first.title #削除されるタスクのタイトル
+        Task.where(user: userid).first.destroy #タスクの削除
+        @linemsg = "bomb!\n\"#{deltitle}\"\nは削除されました！！"
+      when "全削除" #自ユーザのタスクを全て消す
+        Task.where(user: userid).destroy_all #タスクの削除
+        @linemsg = "BooooooomB！\n全てのタスクは削除されました！"
+      when "全表示" #ユーザ関係なく全てのタスク表示
+        list = Task.all.map do |task| #Sp2で修正した。
+          "#{task.id}: #{task.title}"
+        end.join("\n")
+        @linemsg = "【今までに登録したタスク】\n#{list}"
+      else # その他＝タスクの登録
         Task.create!(title: message,user: userid)
         @linemsg = "ぽーーん！\n\"#{message}\"\nが登録されました！！"
     end
